+++ header.php
@@ -1,7 +1,7 @@
 <?
 /**
  * @author      Flurin Dürst
- * @version     1.2.2
+ * @version     1.3
  * @since       WPSeed 0.1
  */
 ?>
@@ -10,7 +10,6 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, minimum-scale=1, maximum-scale=1, initial-scale=1, user-scalable=no">
-    <title><? bloginfo('name') ?><? wp_title('|', true, 'left'); ?></title>
     <link rel="pingback" href="<?php bloginfo( 'pingback_url' ); ?>" />
     <!--===== OPEN-GRAPH TAGS =====-->
     <meta property="og:title" content="<? bloginfo('name'); ?>">
@@ -19,13 +18,7 @@
     <meta property="og:image" content="<? bloginfo('template_url') ?>/dist/images/ogimg.jpg">
     <meta property="og:image:width" content="">
     <meta property="og:image:height" content="">
-
-    <!--===== CSS =====-->
-    <link rel="stylesheet" href="<? bloginfo('template_url') ?>/dist/styles/main.css">
-    <!--===== FONTS =====-->
-    <link href='https://fonts.googleapis.com/css?family=Ubuntu:400,500,700' rel='stylesheet' type='text/css'>
-    <!--===== GOOGLE ANALYTICS =====-->
-    <script>//###</script>
+    <? wp_head(); ?>
   </head>
 
   <body>